+++ src/routes/state/post.state.pages.ep.ts
@@ -1,5 +1,5 @@
 import { FastifyReply, FastifyRequest } from 'fastify'
-import { errr, log_err, task, task_end } from '../../utility/logging'
+import { log_err, task, task_end } from '../../utility/logging'
 import JsonapiErrorBuilder from '../../business.logic/builder/JsonapiErrorBuilder'
 import { default_500_error_response } from '../../business.logic/errors'
 import  { STATE_PAGES, STATE_PAGES_THEME_DARK } from '../../state/page'
@@ -13,32 +13,25 @@ export default async function post_state_pages_endpoint (
   reply: FastifyReply
 ) {
   try {
-    const key = req.body.key
-    const mode = req.body.mode
-    if (!key) {
-      errr(`'key' was not received.`)
+    task('Validating request body... ')
+    if (!req.body.key || !req.body.mode) {
+      task_end('Failed', '❌', '\n[DEBUG][400] Malformed request received.', req.body)
       reply.code(400).send(new JsonapiErrorBuilder()
         .withStatus(400)
-        .withCode('MISSING_DATA')
-        .withTitle('Missing information')
-      )
+        .withCode('MALFORMED_REQUEST')
+        .withTitle('Malformed request received.')
+        .withDetail('Both \'key\' and \'mode\' are required in the request body.')
+        .build())
       return
     }
-    if (!mode) {
-      errr(`'mode' was not received.`)
-      reply.code(400).send(new JsonapiErrorBuilder()
-        .withStatus(400)
-        .withCode('MISSING_DATA')
-        .withTitle('Missing information')
-      )
-      return
-    }
+    const { key, mode } = req.body
+
     task(`Loading '${key}' state... `)
     const light = STATE_PAGES[key]
     const dark = STATE_PAGES_THEME_DARK[key]
     const pageState = themed(light, dark, mode)
     if (pageState) {
-      task_end('Done.')
+      task_end('OK', '✔️')
       reply.code(200).send({
         state: {
           'pages': { [key]: pageState },
@@ -47,7 +40,7 @@ export default async function post_state_pages_endpoint (
         }
       } as TJsonapiStateResponse)
     } else {
-      task_end('Failed.')
+      task_end('Failed.', '❌', `\n[DEBUG][404] Page state for key '${key}' not found.`)
       reply.code(404).send({
         state: {
           'pages': {